Working with precompiled C4lib files
As some of the features we were using in our project weren't yet
supported in the compiled C4lib.a file we were using for the C4
framework, I have migrated the project to a version without the
precompiled library.

Longpress now works.

I have added a method to return the IP Address of the iPad in an
NSString but this method is VERY slow and severly effects performance
thus far. It has been commented out for the time being.